split index.js into multiple files, reuse functions
