+++ remix-tests/src/compiler.js
@@ -1,5 +1,5 @@
 /* eslint no-extend-native: "warn" */
-let fs = require('fs')
+let fs = require('./fs')
 var async = require('async')
 var path = require('path')
 let RemixCompiler = require('remix-solidity').Compiler
@@ -35,58 +35,50 @@ function compileFileOrFiles (filename, isDirectory, opts, cb) {
   }
   // TODO: for now assumes filepath dir contains all tests, later all this
   // should be replaced with remix's & browser solidity compiler code
-  filepath = (isDirectory ? filename : path.dirname(filename))
 
-  // https://github.com/mikeal/node-utils/blob/master/file/lib/main.js
-  fs.walkSync = function (start, callback) {
-    fs.readdirSync(start).forEach(name => {
-      if (name === 'node_modules') {
-        return // hack
-      }
-      var abspath = path.join(start, name)
-      if (fs.statSync(abspath).isDirectory()) {
-        fs.walkSync(abspath, callback)
-      } else {
-        callback(abspath)
+  // This logic is wrong
+  // We should only look into current file if a full file name with path is given
+  // We should only walk through directory if a directory name is passed
+  try {
+    filepath = (isDirectory ? filename : path.dirname(filename))
+    // walkSync only if it is a directory
+    fs.walkSync(filepath, foundpath => {
+      // only process .sol files
+      if (foundpath.split('.').pop() === 'sol') {
+        let c = fs.readFileSync(foundpath).toString()
+        const s = /^(import)\s['"](remix_tests.sol|tests.sol)['"];/gm
+        if (foundpath.indexOf('_test.sol') > 0 && c.regexIndexOf(s) < 0) {
+          c = c.replace(/(pragma solidity \^?\d+\.\d+\.\d+;)/, '$1\nimport \'remix_tests.sol\';')
+        }
+        sources[foundpath] = { content: c }
       }
     })
-  }
-
-  fs.walkSync(filepath, foundpath => {
-    // only process .sol files
-    if (foundpath.split('.').pop() === 'sol') {
-      let c = fs.readFileSync(foundpath).toString()
-      const s = /^(import)\s['"](remix_tests.sol|tests.sol)['"];/gm
-      let includeTestLibs = '\nimport \'remix_tests.sol\';\n'
-      if (file.indexOf('_test.sol') > 0 && c.regexIndexOf(s) < 0) {
-        c = includeTestLibs.concat(c)
+  } catch (e) {
+    throw e
+  } finally {
+    async.waterfall([
+      function loadCompiler (next) {
+        compiler = new RemixCompiler()
+        compiler.onInternalCompilerLoaded()
+        // compiler.event.register('compilerLoaded', this, function (version) {
+        next()
+        // });
+      },
+      function doCompilation (next) {
+        compiler.event.register('compilationFinished', this, function (success, data, source) {
+          next(null, data)
+        })
+        compiler.compile(sources, false)
       }
-      sources[foundpath] = { content: c }
-    }
-  })
-
-  async.waterfall([
-    function loadCompiler (next) {
-      compiler = new RemixCompiler()
-      compiler.onInternalCompilerLoaded()
-      // compiler.event.register('compilerLoaded', this, function (version) {
-      next()
-      // });
-    },
-    function doCompilation (next) {
-      compiler.event.register('compilationFinished', this, function (success, data, source) {
-        next(null, data)
-      })
-      compiler.compile(sources, false)
-    }
-  ], function (err, result) {
-    let errors = (result.errors || []).filter((e) => e.type === 'Error' || e.severity === 'error')
-    if (errors.length > 0) {
-      if (!isBrowser) require('signale').fatal(errors)
-      return cb(new Error('errors compiling'))
-    }
-    cb(err, result.contracts)
-  })
+    ], function (err, result) {
+      let errors = (result.errors || []).filter((e) => e.type === 'Error' || e.severity === 'error')
+      if (errors.length > 0) {
+        if (!isBrowser) require('signale').fatal(errors)
+        return cb(new Error('errors compiling'))
+      }
+      cb(err, result.contracts)
+    })
+  }
 }
 
 function compileContractSources (sources, importFileCb, opts, cb) {
